+++ Attribute/ColorHtmlPropertyAttribute.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace DSC.DialogueSystem
+{
+    public class ColorHtmlPropertyAttribute : PropertyAttribute
+    {
+
+    }
+}